refactor: split monolithic operator_cli.rs into module directory
Split the 1,816-line src/operator_cli.rs into src/operator_cli/ with 9
sub-modules organized by CLI subcommand group:

- mod.rs: top-level dispatch, help text, spawn/handover/bootstrap
- args.rs: shared arg parsing helpers (next_required, reject_extra_args, etc.)
- engineer.rs: engineer subcommands
- meeting.rs: meeting subcommands
- curation.rs: goal-curation + improvement-curation subcommands
- review.rs: review subcommands
- gym.rs: gym subcommands
- ooda.rs: OODA subcommands
- decisions.rs: act-on-decisions + GitHub issue creation

All 96 operator_cli tests pass. Tests kept with their related code.
No public API changes.
